get synchronizedAddMessages tests running again
