+++ .eslintrc.yaml
@@ -177,7 +177,7 @@ overrides:
       - "packages/studio-base/src/panels/Image/components/Toolbar.tsx"
       - "packages/studio-base/src/panels/Image/components/TopicDropdown.tsx"
       - "packages/studio-base/src/panels/Image/components/ZoomMenu.tsx"
-      - "packages/studio-base/src/panels/Image/index.tsx"
+      - "packages/studio-base/src/panels/Image/ImageView.tsx"
       - "packages/studio-base/src/panels/NodePlayground/BottomBar/DiagnosticsSection.tsx"
       - "packages/studio-base/src/panels/NodePlayground/BottomBar/index.tsx"
       - "packages/studio-base/src/panels/NodePlayground/Sidebar.tsx"
