+++ packages/studio-base/src/panels/Image/ImageView.tsx
@@ -35,7 +35,6 @@ import { formatTimeRaw } from "@foxglove/studio-base/util/time";
 
 import { ImageCanvas, ImageEmptyState, Toolbar, TopicDropdown } from "./components";
 import { useCameraInfo, ANNOTATION_DATATYPES, useImagePanelMessages } from "./hooks";
-import helpContent from "./index.help.md";
 import { downloadImage } from "./lib/downloadImage";
 import { NORMALIZABLE_IMAGE_DATATYPES } from "./lib/normalizeMessage";
 import { getRelatedMarkerTopics, getMarkerOptions } from "./lib/util";
@@ -71,7 +70,9 @@ export function ImageView({ context }: Props): JSX.Element {
   const [renderDone, setRenderDone] = useState(() => () => {});
   const [topics, setTopics] = useState<readonly Topic[]>([]);
   const [config, setConfig] = useState<Config>(() => {
-    const initialConfig = context.initialState as DeepPartial<Config>;
+    const initialConfig = context.initialState as DeepPartial<Config> & {
+      enabledMarkerTopics?: Config["enabledMarkerTopics"];
+    };
     return {
       ...defaultConfig,
       ...initialConfig,
@@ -311,7 +312,7 @@ export function ImageView({ context }: Props): JSX.Element {
 
   return (
     <Stack flex="auto" overflow="hidden" position="relative">
-      <PanelToolbar helpContent={helpContent}>
+      <PanelToolbar>
         <Stack direction="row" flex="auto" alignItems="center" overflow="hidden">
           <TopicDropdown
             topics={imageTopics}
