+++ packages/studio-base/src/panels/Image/hooks/useImagePanelMessages.ts
@@ -6,20 +6,15 @@ import { useState } from "react";
 import { createStore, useStore } from "zustand";
 
 import { AVLTree } from "@foxglove/avl";
-import {
-  Time,
-  compare as compareTime,
-  isLessThan,
-  toNanoSec,
-  fromNanoSec,
-} from "@foxglove/rostime";
+import { Time, compare as compareTime } from "@foxglove/rostime";
 import { MessageEvent, RenderState } from "@foxglove/studio";
 
 import { normalizeAnnotations } from "../lib/normalizeAnnotations";
 import { normalizeImageMessage } from "../lib/normalizeMessage";
 import { Annotation, NormalizedImageMessage } from "../types";
+import { synchronizedAddMessages } from "./synchronizedAddMessages";
 
-type ImagePanelState = {
+export type ImagePanelState = {
   imageTopic?: string;
   annotationTopics: string[];
   image?: NormalizedImageMessage;
@@ -28,7 +23,7 @@ type ImagePanelState = {
   synchronize: boolean;
 
   actions: {
-    frame(currentFrame: NonNullable<RenderState["currentFrame"]>): void;
+    setCurrentFrame(currentFrame: NonNullable<RenderState["currentFrame"]>): void;
     clear(): void;
     setSynchronize(synchronize: boolean): void;
     setImageTopic(newTopic: string): void;
@@ -73,92 +68,6 @@ export const ANNOTATION_DATATYPES = [
   "foxglove.ImageAnnotations",
 ] as const;
 
-export function synchronizedAddMessages(
-  state: ImagePanelState,
-  messageEvents: readonly MessageEvent<unknown>[],
-): Partial<ImagePanelState> {
-  for (const messageEvent of messageEvents) {
-    const image = normalizeImageMessage(messageEvent, messageEvent.schemaName);
-    const annotations = normalizeAnnotations(messageEvent, messageEvent.schemaName);
-    if (!image && !annotations) {
-      continue;
-    }
-
-    if (image) {
-      // Update the image at the stamp time
-      const item = state.tree.get(image.stamp);
-      if (item) {
-        item.image = image;
-      } else {
-        state.tree.set(image.stamp, {
-          image,
-          annotationsByTopic: new Map(),
-        });
-      }
-    }
-
-    if (annotations) {
-      // Group annotations by timestamp, then update the annotations by topic at each stamp
-      const groups = new Map<bigint, Annotation[]>();
-      for (const annotation of annotations) {
-        const key = toNanoSec(annotation.stamp);
-        const arr = groups.get(key);
-        if (arr) {
-          arr.push(annotation);
-        } else {
-          groups.set(key, [annotation]);
-        }
-      }
-
-      for (const [stampNsec, group] of groups) {
-        const stamp = fromNanoSec(stampNsec);
-        let item = state.tree.get(stamp);
-        if (!item) {
-          item = {
-            image: undefined,
-            annotationsByTopic: new Map(),
-          };
-          state.tree.set(stamp, item);
-        }
-        item.annotationsByTopic.set(messageEvent.topic, group);
-      }
-    }
-
-    // Find the oldest entry where we have everything synchronized
-    let validEntry: [Time, SynchronizationItem] | undefined = undefined;
-    for (const entry of state.tree.entries()) {
-      const messageState = entry[1];
-      // If we have an image and all the messages for annotation topics then we have a synchronized set.
-      if (
-        messageState.image &&
-        messageState.annotationsByTopic.size === state.annotationTopics.length
-      ) {
-        validEntry = entry;
-      }
-    }
-
-    // We've got a set of synchronized messages, remove any older items from the tree
-    if (validEntry) {
-      let minKey = state.tree.minKey();
-      while (minKey && isLessThan(minKey, validEntry[0])) {
-        state.tree.shift();
-
-        minKey = state.tree.minKey();
-      }
-
-      return {
-        imageTopic: state.imageTopic,
-        image: validEntry[1].image,
-        annotationsByTopic: validEntry[1].annotationsByTopic,
-        tree: state.tree,
-      };
-    }
-  }
-
-  // with no valid entry, we keep the previous state
-  return state;
-}
-
 function addMessages(
   state: ImagePanelState,
   messageEvents: readonly MessageEvent<unknown>[],
@@ -193,7 +102,7 @@ function addMessages(
   return newState ?? state;
 }
 
-function useImagePanelMessages(initialState: {
+export function useImagePanelMessages(initialState: {
   imageTopic: string;
   annotationTopics: string[];
   synchronize: boolean;
@@ -236,7 +145,7 @@ function useImagePanelMessages(initialState: {
             tree: new AVLTree(compareTime),
           });
         },
-        frame(currentFrame) {
+        setCurrentFrame(currentFrame) {
           set((old) => addMessages(old, currentFrame));
         },
       },
@@ -245,5 +154,3 @@ function useImagePanelMessages(initialState: {
 
   return useStore(store, selectPublicState);
 }
-
-export { useImagePanelMessages };
