+++ packages/studio-base/src/panels/Image/types.ts
@@ -11,8 +11,6 @@ export type DefaultConfig = {
   cameraTopic: string;
   enabledMarkerTopics: string[];
   synchronize: boolean;
-  /** @deprecated */
-  customMarkerTopicOptions?: string[];
 };
 
 export type Config = DefaultConfig & {
